model/clip_server/requirements.txt
fastapi
uvicorn
torch
transformers
Pillow
